nothing special

+++ l1/src/main.cpp
@@ -1,7 +1,7 @@
 #include "matrix.hpp"
 #include <iostream>
 int main(){
-    std::vector<std::vector<double>> matr;
+    std::vector<std::vector<int>> matr;
     init_matr(matr); 
     std::cout << matr;
 }
